Reorganize alerts
