+++ context/app/static/js/components/detailPage/alerts/ExternalDatasetAlert/ExternalDatasetAlert.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import { Alert } from 'js/shared-styles/alerts';
+
+function ExternalDatasetAlert() {
+  return (
+    <Alert severity="info" $marginBottom="16">
+      You are viewing an external dataset that was not generated by the HuBMAP Consortium.
+    </Alert>
+  );
+}
+
+export default ExternalDatasetAlert;
