+++ context/app/static/js/components/detailPage/alerts/SharedAlerts/SharedAlerts.jsx
