+++ context/app/static/js/components/detailPage/alerts/SharedAlerts/index.js
