+++ context/app/static/js/pages/Dataset/Dataset.jsx
@@ -13,10 +13,11 @@ import SummaryItem from 'js/components/detailPage/summary/SummaryItem';
 import ContributorsTable from 'js/components/detailPage/ContributorsTable';
 import useSendUUIDEvent from 'js/components/detailPage/useSendUUIDEvent';
 import CollectionsSection from 'js/components/detailPage/CollectionsSection';
-import SupportAlert from 'js/components/detailPage/SupportAlert';
 import { useSearchHits } from 'js/hooks/useSearchData';
 import { getAllCollectionsQuery } from 'js/helpers/queries';
-import { Alert } from 'js/shared-styles/alerts';
+import SupportAlert from 'js/components/detailPage/alerts/SupportAlert';
+import ExternalDatasetAlert from 'js/components/detailPage/alerts/ExternalDatasetAlert/ExternalDatasetAlert';
+import VersionAlert from 'js/components/detailPage/alerts/VersionAlert/VersionAlert';
 
 // TODO use this context for components other than FileBrowser
 import DetailContext from 'js/components/detailPage/context';
@@ -119,20 +120,8 @@ function DatasetDetail(props) {
         sectionOrder={sectionOrder}
         additionalAlerts={
           <>
-            {next_revision_uuid && (
-              <Alert severity="warning" $marginBottom="16">
-                <span>
-                  {/* <span> to override "display: flex" which splits this on to multiple lines. */}
-                  You are viewing an older version of this page. Navigate to the{' '}
-                  <LightBlueLink href={`/browse/latest/dataset/${uuid}`}>latest version</LightBlueLink>.
-                </span>
-              </Alert>
-            )}
-            {['Dataset', 'Support'].includes(mapped_external_group_name) && (
-              <Alert severity="info" $marginBottom="16">
-                You are viewing an external dataset that was not generated by the HuBMAP Consortium.
-              </Alert>
-            )}
+            {next_revision_uuid && <VersionAlert uuid={uuid} />}
+            {['Dataset', 'Support'].includes(mapped_external_group_name) && <ExternalDatasetAlert />}
             {entity_type === 'Support' && <SupportAlert uuid={uuid} $marginBottom="16" />}
           </>
         }
